Final neatening and Script Demo
While I'm not so thrilled with the 'amount' of tests I wrote, this was an interesting little project.
I enjoyed putting together a process that both met your guides and challenged me. I could have just as easily built this
by hard coding ten different functions and a hash of operators. Probably alot easier really. But if you want scalability and
re-usability a module is the way to go, I was halfway through build a gem for it at one point honestly but thats a new area
for me so I figured I'd stick to the easy stuff.